+++ api/controller/brands.go
@@ -1,4 +1,4 @@
-package api
+package controller
 
 import (
 	"encoding/json"
